:recycle: Refactor custom matrix code into submodule
