+++ render-wasm/src/shapes/renderable.rs
@@ -7,21 +7,7 @@ use crate::render::{ImageStore, Renderable};
 
 impl Renderable for Shape {
     fn render(&self, surface: &mut skia_safe::Surface, images: &ImageStore) -> Result<(), String> {
-        let mut transform = skia::Matrix::new_identity();
-        let (translate_x, translate_y) = self.translation();
-        let (scale_x, scale_y) = self.scale();
-        let (skew_x, skew_y) = self.skew();
-        transform.set_all(
-            scale_x,
-            skew_x,
-            translate_x,
-            skew_y,
-            scale_y,
-            translate_y,
-            0.,
-            0.,
-            1.,
-        );
+        let transform = self.transform.to_skia_matrix();
 
         // Check transform-matrix code from common/src/app/common/geom/shapes/transforms.cljc
         let center = self.selrect.center();
